fix(specpart): convert input at the C boundary; make kernel reentrant
Fixes GH-142. The C wrapper read the input buffer via PyArray_DATA with
no dtype/contiguity checks, silently mis-reading any transposed or
strided input. Datasets stored with dims (..., dir, freq) - like the
one attached to the issue - are transposed to (freq, dir) by the
partition pipeline, producing an F-ordered view that astype(float32)
(order='K') preserves; the kernel then read the bytes as C-ordered and
returned structurally wrong partitions (up to 1.59 m Hs error on the
issue dataset). The f2py wrapper around the original Fortran code
handled strides correctly, which is why results changed from v3 to v4.

The wrapper now converts any input to an aligned C-contiguous float32
array (PyArray_FROM_OTF, matching f2py's permissive behaviour),
validates ndim/ihmax, raises instead of exit()ing, and releases the
GIL during computation. The kernel is rewritten without global state
(reentrant, safe under dask threads), with allocation checks and
without the neigh double-malloc leak. The redundant .astype(np.float32)
at the call sites is dropped; conversion is owned by the wrapper.

Verified bit-identical partition maps against the pre-v4 Fortran
reference in 1420 comparisons (real model output, the issue-142
dataset, synthetic/tie-stressed/degenerate spectra x ihmax
25/50/100/200), plus layout/dtype invariance for all 234 issue-142
spectra and 2000-call threaded consistency. Full analysis in
wavespectra/partition/specpart/ANALYSIS.md.

+++ wavespectra/partition/partition.py
@@ -830,7 +830,7 @@ def np_ptm1(
 
     """
     # Use smooth spectrum to define morphological boundaries
-    watershed_map = specpart.partition(spectrum_smooth.astype(np.float32), ihmax)
+    watershed_map = specpart.partition(spectrum_smooth, ihmax)
     nparts = watershed_map.max()
 
     # Wind sea mask
@@ -910,7 +910,7 @@ def np_ptm2(
 
     """
     # Use smooth spectrum to define morphological boundaries
-    watershed_map = specpart.partition(spectrum_smooth.astype(np.float32), ihmax)
+    watershed_map = specpart.partition(spectrum_smooth, ihmax)
     nparts = watershed_map.max()
 
     # Wind sea mask
@@ -980,7 +980,7 @@ def np_ptm3(
 
     """
     # Use smooth spectrum to define morphological boundaries
-    watershed_map = specpart.partition(spectrum_smooth.astype(np.float32), ihmax)
+    watershed_map = specpart.partition(spectrum_smooth, ihmax)
     nparts = watershed_map.max()
 
     # Assign partitioned arrays from raw spectrum and morphological boundaries
@@ -1052,7 +1052,7 @@ def np_hp01(
 
     """
     # Use smooth spectrum to define morphological boundaries
-    watershed_map = specpart.partition(spectrum_smooth.astype(np.float32), ihmax)
+    watershed_map = specpart.partition(spectrum_smooth, ihmax)
     nparts = watershed_map.max()
 
     # Assign partitioned arrays from raw spectrum and morphological boundaries
@@ -1145,7 +1145,7 @@ def np_hp01_wseabins(
 
     """
     # Use smooth spectrum to define morphological boundaries
-    watershed_map = specpart.partition(spectrum_smooth.astype(np.float32), ihmax)
+    watershed_map = specpart.partition(spectrum_smooth, ihmax)
     nparts = watershed_map.max()
 
     # Assign partitioned arrays from raw spectrum and morphological boundaries
@@ -1241,7 +1241,7 @@ def np_hp01_wseafrac_wseabins(
 
     """
     # Use smooth spectrum to define morphological boundaries
-    watershed_map = specpart.partition(spectrum_smooth.astype(np.float32), ihmax)
+    watershed_map = specpart.partition(spectrum_smooth, ihmax)
     nparts = watershed_map.max()
 
     # Assign partitioned arrays from raw spectrum and morphological boundaries
